LPS-195225 Change SourceDataRenderer to be reusable
