+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/FDSPropsTransformer/MultiselectPicklistFDSPropsTransformer.tsx
@@ -3,8 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import classNames from 'classnames';
-import React from 'react';
+import SourceDataRenderer from './FDSDataRenderers/SourceDataRenderer';
 
 interface MultiselectPicklistDataRendererProps {
 	value: {
@@ -12,25 +11,6 @@ interface MultiselectPicklistDataRendererProps {
 		name: string;
 		name_i18n: string;
 	}[];
-}
-
-interface SourceDataRenderer {
-	value: boolean;
-}
-
-function SourceDataRenderer({value}: SourceDataRenderer) {
-	return (
-		<strong
-			className={classNames(
-				value ? 'label-info' : 'label-warning',
-				'label'
-			)}
-		>
-			{value
-				? Liferay.Language.get('system')
-				: Liferay.Language.get('custom')}
-		</strong>
-	);
 }
 
 export default function MultiselectPicklistFDSPropsTransformer({
