+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/FDSPropsTransformer/ObjectFieldsFDSPropsTransformer.tsx
@@ -3,33 +3,13 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import classNames from 'classnames';
-import React from 'react';
-
-interface ObjectFieldSourceDataRenderer {
-	value: boolean;
-}
-
-function ObjectFieldSourceDataRenderer({value}: ObjectFieldSourceDataRenderer) {
-	return (
-		<strong
-			className={classNames(
-				value ? 'label-info' : 'label-warning',
-				'label'
-			)}
-		>
-			{value
-				? Liferay.Language.get('system')
-				: Liferay.Language.get('custom')}
-		</strong>
-	);
-}
+import SourceDataRenderer from './FDSDataRenderers/SourceDataRenderer';
 
 export default function ObjectFieldsFDSPropsTransformer({...otherProps}) {
 	return {
 		...otherProps,
 		customDataRenderers: {
-			objectFieldSourceDataRenderer: ObjectFieldSourceDataRenderer,
+			objectFieldSourceDataRenderer: SourceDataRenderer,
 		},
 		onActionDropdownItemClick({
 			action,
