+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/FDSPropsTransformer/FDSDataRenderers/SourceDataRenderer.d.ts
@@ -1,0 +1,14 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+/// <reference types="react" />
+
+interface SourceDataRenderer {
+	value: boolean;
+}
+export default function SourceDataRenderer({
+	value,
+}: SourceDataRenderer): JSX.Element;
+export {};
