+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/FDSPropsTransformer/MultiselectPicklistFDSPropsTransformer.d.ts
@@ -3,8 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-/// <reference types="react" />
-
+import SourceDataRenderer from './FDSDataRenderers/SourceDataRenderer';
 interface MultiselectPicklistDataRendererProps {
 	value: {
 		key: string;
@@ -12,10 +11,6 @@ interface MultiselectPicklistDataRendererProps {
 		name_i18n: string;
 	}[];
 }
-interface SourceDataRenderer {
-	value: boolean;
-}
-declare function SourceDataRenderer({value}: SourceDataRenderer): JSX.Element;
 export default function MultiselectPicklistFDSPropsTransformer({
 	...otherProps
 }: {
