+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/FDSPropsTransformer/ObjectFieldsFDSPropsTransformer.d.ts
@@ -3,21 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-/// <reference types="react" />
-
-interface ObjectFieldSourceDataRenderer {
-	value: boolean;
-}
-declare function ObjectFieldSourceDataRenderer({
-	value,
-}: ObjectFieldSourceDataRenderer): JSX.Element;
+import SourceDataRenderer from './FDSDataRenderers/SourceDataRenderer';
 export default function ObjectFieldsFDSPropsTransformer({
 	...otherProps
 }: {
 	[x: string]: any;
 }): {
 	customDataRenderers: {
-		objectFieldSourceDataRenderer: typeof ObjectFieldSourceDataRenderer;
+		objectFieldSourceDataRenderer: typeof SourceDataRenderer;
 	};
 	onActionDropdownItemClick({
 		action,
@@ -33,4 +26,3 @@ export default function ObjectFieldsFDSPropsTransformer({
 		};
 	}): void;
 };
-export {};
